refactor(bayn): add pure strategy composition boundary
