+++ services/bayn/src/application-plan.ts
@@ -13,7 +13,8 @@ import {
 import { operationalError } from './errors'
 import { canonicalHashV1Result, type CanonicalJsonFailure } from './hash'
 import { loadDefaultProtocol, type CausalProtocol } from './protocol'
-import { makeStrategy, type Strategy } from './strategy'
+import { type Strategy } from './strategy'
+import { makeRiskBalancedTrendStrategy } from './strategy/risk-balanced-trend/strategy'
 
 type RuntimeIdentityFailure =
   | {
@@ -37,6 +38,9 @@ type RuntimeSeed = {
 type ParameterizedRuntime = RuntimeSeed & { readonly parameterHash: string }
 type ProvenanceRuntime = ParameterizedRuntime & { readonly provenance: RuntimeProvenance }
 type StrategyRuntime = ProvenanceRuntime & { readonly strategy: Strategy }
+
+const selectStrategy = (runtime: ProvenanceRuntime): Strategy =>
+  makeRiskBalancedTrendStrategy(runtime.protocol, runtime.provenance)
 
 const hashRuntimeParameters = (seed: RuntimeSeed): Result.Result<ParameterizedRuntime, RuntimeIdentityFailure> =>
   pipe(
@@ -71,10 +75,7 @@ const addRuntimeProvenance = (
     Result.map((provenance) => ({ ...parameterized, provenance })),
   )
 
-const addStrategy = (runtime: ProvenanceRuntime): StrategyRuntime => ({
-  ...runtime,
-  strategy: makeStrategy(runtime.protocol, runtime.provenance),
-})
+const addStrategy = (runtime: ProvenanceRuntime): StrategyRuntime => ({ ...runtime, strategy: selectStrategy(runtime) })
 
 const addStrategyProtocolHash = (
   runtime: StrategyRuntime,
