+++ services/bayn/src/strategy/risk-balanced-trend/index.ts
@@ -1,0 +1,9 @@
+export { makeRiskBalancedTrendDefinition } from './decision'
+export type {
+  RiskBalancedTrendMarketContext,
+  RiskBalancedTrendPortfolioContext,
+  RiskBalancedTrendStrategyDefinition,
+  RiskBalancedTrendTargetPortfolio,
+} from './decision'
+export { makeRiskBalancedTrendStrategy } from './strategy'
+export type { RiskBalancedTrendStrategy, RiskBalancedTrendStrategyPrepareLockFailure } from './strategy'
